Revert "[Add] box Collider componaent"
This reverts commit 012fb223f280bc4a80415099d49440b0db52fafa.

+++ AbsoluteFate/Texture.cpp
@@ -1,4 +1,5 @@
 #include "Texture.h"
+#include "SDL_image.h"
 
 bool Texture::LoadTexture(Renderer& pRenderer, const string& filename)
 {
@@ -14,7 +15,7 @@ bool Texture::LoadTexture(Renderer& pRenderer, const string& filename)
 	mTextureHeight = surface->h;
 
 	//Create texture from surface
-	mSdlTexture = SDL_CreateTextureFromSurface(/**/pRenderer.GetSdlRenderer(), surface); // get mSdlRenderer to Renderer
+	mSdlTexture = SDL_CreateTextureFromSurface(pRenderer.GetSdlRenderer(), surface); // get mSdlRenderer to Renderer
 	SDL_FreeSurface(surface);
 	if (!mSdlTexture)
 	{
@@ -53,5 +54,5 @@ int Texture::GetHeight()
 
 Vector2D Texture::GetSize()
 {
-	return Vector2D((float)mTextureHeight, (float)mTextureHeight);
+	return Vector2D(mTextureHeight, mTextureHeight);
 }
